wip: refactor in progress
